# [NVIDIA] Support Flashinfer TRTLLM FP8-q/kv NVFP4-out Attention Kernel