--- a/vllm/attention/backends/flash_attn.py
+++ b/vllm/attention/backends/flash_attn.py
@@ -509,11 +509,11 @@ class FlashAttentionMetadataBuilder(
 class FlashAttentionImpl(AttentionImpl):
     """
     If the input tensors contain prompt tokens, the layout is as follows:
-    |<--------------- num_prefill_tokens ----------------->|
+    |<--------------- num_prefill_tokens ----------------->|	
     |<--prefill_0-->|<--prefill_1-->|...|<--prefill_N-1--->|
 
-    Otherwise, the layout is as follows:
-    |<----------------- num_decode_tokens ------------------>|
+    Otherwise, the layout is as follows:	
+    |<----------------- num_decode_tokens ------------------>|	
     |<--decode_0-->|..........|<--decode_M-1-->|<--padding-->|
 
     Generation tokens can contain padding when cuda-graph is used.
@@ -855,10 +855,10 @@ def _get_query_key_seq_metadata(
     attn_type: str,
 ) -> tuple:
     """
-    Returns sequence metadata for key and query based on the specified
+    Returns sequence metadata for key and query based on the specified 
     attention type and whether input is a prompt.
 
-    This function computes the starting locations and maximum sequence lengths
+    This function computes the starting locations and maximum sequence lengths 
     for key and query sequences for different attention types.
 
     Args:
@@ -916,15 +916,15 @@ def _get_query_key_seq_metadata(
 
 def _get_causal_option(attn_type: str) -> bool:
     """
-    Determine whether the given attention type is suitable for causal
+    Determine whether the given attention type is suitable for causal 
     attention mechanisms.
 
     Args:
         attn_type (AttentionType): The type of attention being evaluated
 
     Returns:
-        bool: Returns `True` if the attention type is suitable for causal
-        attention (i.e., not encoder, encoder-only, or encoder-decoder),
+        bool: Returns `True` if the attention type is suitable for causal 
+        attention (i.e., not encoder, encoder-only, or encoder-decoder), 
         otherwise returns `False`.
     """
     return not (attn_type == AttentionType.ENCODER