# Essential Elements of an Effective PR Description Checklist

- [x] The purpose of the PR, such as "Fix some issue (link existing issues this PR will resolve)".
- [x] The test plan, such as providing test command.
- [x] The test results, such as pasting the results comparison before and after, or e2e results
- [ ] (Optional) The necessary documentation update, such as updating `supported_models.md` and `examples` for a new model.


## Purpose
This PR based on the previous attn + FP8-quant fusion #21716, adding another attn + NVFP4-quant fusion for supporting TRTLLM-gen attn kernel.


## Test Plan && Test Result
Functional:
- NVFP4 TRTLLM Prefill/Decode kernel unit test: tests/kernels/attention/test_flashinfer_trtllm_attention.py
```
======= 44 passed, 4 skipped, 4 warnings in 64.13s (0:01:04) ======
```
- AttentionStaticQuantPattern fusion_attn pass: `tests/compile/test_fusion_attn.py::test_attention_quant_pattern`
```
======== 12 passed, 4 warnings in 68.53s (0:01:08) ========
```

E2E Performance: `nvidia/Llama-4-Scout-17B-16E-Instruct-FP4`

main
```
--kv-cache-dtype = auto                                                      --kv-cache-dtype = fp8
============ Serving Benchmark Result ============                           ============ Serving Benchmark Result ============
Successful requests:                     640                                 Successful requests:                     640
Maximum request concurrency:             128                                 Maximum request concurrency:             128
Benchmark duration (s):                  143.36                              Benchmark duration (s):                  164.40
Total input tokens:                      653975                              Total input tokens:                      653975
Total generated tokens:                  655360                              Total generated tokens:                  655360
Request throughput (req/s):              4.46                                Request throughput (req/s):              3.89
Output token throughput (tok/s):         4571.39                             Output token throughput (tok/s):         3986.50
Total Token throughput (tok/s):          9133.12                             Total Token throughput (tok/s):          7964.57
---------------Time to First Token----------------                           ---------------Time to First Token----------------
Mean TTFT (ms):                          677.91                              Mean TTFT (ms):                          751.94
Median TTFT (ms):                        475.19                              Median TTFT (ms):                        521.30
P99 TTFT (ms):                           2509.39                             P99 TTFT (ms):                           2859.41
-----Time per Output Token (excl. 1st token)------                           -----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          27.32                               Mean TPOT (ms):                          31.36
Median TPOT (ms):                        27.32                               Median TPOT (ms):                        31.40
P99 TPOT (ms):                           28.20                               P99 TPOT (ms):                           32.38
---------------Inter-token Latency----------------                           ---------------Inter-token Latency----------------
Mean ITL (ms):                           27.32                               Mean ITL (ms):                           31.36
Median ITL (ms):                         25.66                               Median ITL (ms):                         29.73
P99 ITL (ms):                            146.45                              P99 ITL (ms):                            168.08
==================================================                           ==================================================
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|    |Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|    |-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.9037|±  |0.0081|    |gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.8969|±  |0.0084|
|     |       |strict-match    |     5|exact_match|↑  |0.8908|±  |0.0086|    |     |       |strict-match    |     5|exact_match|↑  |0.8817|±  |0.0089|
```

PR
```
--kv-cache-dtype = auto                                                      --kv-cache-dtype = fp8
============ Serving Benchmark Result ============                           ============ Serving Benchmark Result ============
Successful requests:                     640                                 Successful requests:                     640
Maximum request concurrency:             128                                 Maximum request concurrency:             128
Benchmark duration (s):                  143.82                              Benchmark duration (s):                  164.89
Total input tokens:                      653975                              Total input tokens:                      653975
Total generated tokens:                  655360                              Total generated tokens:                  655360
Request throughput (req/s):              4.45                                Request throughput (req/s):              3.88
Output token throughput (tok/s):         4556.72                             Output token throughput (tok/s):         3974.57
Total Token throughput (tok/s):          9103.80                             Total Token throughput (tok/s):          7940.75
---------------Time to First Token----------------                           ---------------Time to First Token----------------
Mean TTFT (ms):                          694.81                              Mean TTFT (ms):                          750.19
Median TTFT (ms):                        476.66                              Median TTFT (ms):                        522.09
P99 TTFT (ms):                           2595.24                             P99 TTFT (ms):                           2847.96
-----Time per Output Token (excl. 1st token)------                           -----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          27.40                               Mean TPOT (ms):                          31.46
Median TPOT (ms):                        27.38                               Median TPOT (ms):                        31.51
P99 TPOT (ms):                           28.52                               P99 TPOT (ms):                           32.46
---------------Inter-token Latency----------------                           ---------------Inter-token Latency----------------
Mean ITL (ms):                           27.40                               Mean ITL (ms):                           31.46
Median ITL (ms):                         25.68                               Median ITL (ms):                         29.84
P99 ITL (ms):                            153.03                              P99 ITL (ms):                            171.23
==================================================                           ==================================================
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|    |Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|    |-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.9075|±  |0.0080|    |gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.8984|±  |0.0083|
|     |       |strict-match    |     5|exact_match|↑  |0.8954|±  |0.0084|    |     |       |strict-match    |     5|exact_match|↑  |0.8832|±  |0.0088|

--kv-cache-dtype = fp8, enable_attn_fusion
============ Serving Benchmark Result ============
Successful requests:                     640
Maximum request concurrency:             128
Benchmark duration (s):                  130.05
Total input tokens:                      653975
Total generated tokens:                  655360
Request throughput (req/s):              4.92
Output token throughput (tok/s):         5039.38
Total Token throughput (tok/s):          10068.10
---------------Time to First Token----------------
Mean TTFT (ms):                          658.54
Median TTFT (ms):                        456.26
P99 TTFT (ms):                           2400.84
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          24.74
Median TPOT (ms):                        24.74
P99 TPOT (ms):                           25.72
---------------Inter-token Latency----------------
Mean ITL (ms):                           24.74
Median ITL (ms):                         23.13
P99 ITL (ms):                            140.43
==================================================
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.9014|±  |0.0082|
|     |       |strict-match    |     5|exact_match|↑  |0.8863|±  |0.0087|
```

## (Optional) Documentation Update