@nvpohanh 2025-08-19
@elvischenv Could you rebase this PR on top of current https://github.com/vllm-project/vllm/pull/21716 ? Thanks!