@ProExpertProg 2025-08-21
Yeah I think you're importing fusion passes into attention which we shouldn't do in general. Can you extract QuantKey (but not QUANT_OPS) into the same place as GroupShape?