@ProExpertProg 2025-08-21
Also QuantKey should be frozen (immutable) if not done yet